Issue #62: Fixed NavigateToPageAction to find INavigate ancestors
The problem was NavigateToPageAction's management of sender vs.
senderObject in Execute's loop looking for an INavigate implementation.
Both implementations (managed and native) had similar problems where the
current element's visual parent was found, but wasn't actually analyzed on
the next time around the loop.

The fixes in the Execute methods are tiny.  The rest of the changes are
all about unit testing the managed fix.  I didn't invest in unit tests for
the native change, but did verify it manually in a test app.