Fixed #76: Properly check for Windows Runtime events in EventTriggerBehavior
Developers can reuse the Windows Runtime event handlers in their custom
controls, but for those cases, the EventTriggerBerhavior shouldn't use
the WindowsRuntimeMarshal class.

This change ensures that not only the EventInfo.EventHandler is checked
for Windows Runtime type, but also that the EventInfo.DeclaringType is
also checked.

This ensures that a custom button extendind the Button Windows Runtime
class will still consider the Click event as a Windows Runtime event,
but a custom MyClick event in it, wouldn't.