Fixed #89: Ensured EventTriggerBehavior supports x:Bind bindings
For x:Bind to work with EventTriggerbehavior, we need to ensure that no
exception gets raised for not finding the request event on the source
object, as this will might (will) get evaluate at a later phase.

So instead of raising an exception, we just gracefully return from the
method.