Fixed #103: Check for CommandParamenter value set instead of non-null value
As it currently stands, the InvokeCommandAction.CommandParamenter can
never have a null value as it will make it fallback to the other
approaches, which can raise an InvalidCastException.

A null value should be allowed for the CommandParameter, so instead of
doing a null check, I've replaced it with a property unset check
instead.